♻[견적서페이지]: 타입 위치 수정
